Conflicts:
	TFC_Shared/src/TFC/Blocks/Vanilla/BlockCustomIce.java
	TFC_Shared/src/TFC/TFCItems.java
	src/Common/com/bioxx/tfc/Core/TFC_Core.java
	src/Common/com/bioxx/tfc/Handlers/EntityDamageHandler.java
	src/Common/com/bioxx/tfc/Items/Tools/ItemCustomBow.java
	src/Common/com/bioxx/tfc/Reference.java
	src/Common/com/bioxx/tfc/TileEntities/TEFoodPrep.java

+++ src/API/com/bioxx/tfc/api/Constant/Global.java
@@ -116,6 +116,6 @@ public class Global
 	 * environmental factors, food will decay at 50% per 24 hours.
 	 * Easy calculator here: http://www.basic-mathematics.com/nth-root-calculator.html
 	 */
-	public static float FOOD_DECAY_RATE = 1.0170378966055869517978300569768f;
+	public static double FOOD_DECAY_RATE = 1.0170378966055869517978300569768f;
 	public static float FOOD_MAX_WEIGHT = 160;
 }
